Refactoring UI to allow root views to include other list and detail views
